1. change type of Ask, Bid(string -> int64) 2. change command function 3. change CollectStockLog Function

+++ src/stockbot/model/trademodel.go
@@ -10,7 +10,7 @@ type TradeLog struct {
 	StockCode   string
 	StockMarket string
 	TradeType   string
-	Price       uint
+	Price       int64
 	Count       uint
 	Timestamp   time.Time
 }
